removing deprecated methods that call the voice, application, number insight, sms and verify apis
