+++ tests/test_sms.py
@@ -53,45 +53,6 @@ def test_submit_sms_conversion(sms):
 
 
 @responses.activate
-def test_deprecated_send_message(client, dummy_data):
-    stub(responses.POST, "https://rest.nexmo.com/sms/json")
-
-    params = {"from": "Python", "to": "447525856424", "text": "Hey!"}
-
-    assert isinstance(client.send_message(params), dict)
-    assert request_user_agent() == dummy_data.user_agent
-    assert "from=Python" in request_body()
-    assert "to=447525856424" in request_body()
-    assert "text=Hey%21" in request_body()
-
-
-@responses.activate
-def test_deprecated_authentication_error(client):
-    responses.add(responses.POST, "https://rest.nexmo.com/sms/json", status=401)
-
-    with pytest.raises(vonage.AuthenticationError):
-        client.send_message({})
-
-
-@responses.activate
-def test_deprecated_client_error(client):
-    responses.add(responses.POST, "https://rest.nexmo.com/sms/json", status=400)
-
-    with pytest.raises(vonage.ClientError) as excinfo:
-        client.send_message({})
-    excinfo.match(r"400 response from rest.nexmo.com")
-
-
-@responses.activate
-def test_deprecated_server_error(client):
-    responses.add(responses.POST, "https://rest.nexmo.com/sms/json", status=500)
-
-    with pytest.raises(vonage.ServerError) as excinfo:
-        client.send_message({})
-    excinfo.match(r"500 response from rest.nexmo.com")
-
-
-@responses.activate
 def test_deprecated_submit_sms_conversion(client):
     responses.add(
         responses.POST, "https://api.nexmo.com/conversions/sms", status=200, body=b"OK"
